+++ commands/default_cmdsets.py
@@ -43,6 +43,7 @@ from features.ranks.commands import RankCmdSet
 from features.professions.commands import ProfessionCmdSet
 from features.dungeons.commands import DungeonCmdSet
 from features.trade.commands import TradeCmdSet
+from features.mail.commands import MailCmdSet
 from commands.general_commands import GeneralCmdSet
 
 
@@ -118,6 +119,8 @@ class CharacterCmdSet(default_cmds.CharacterCmdSet):
         self.add(DungeonCmdSet)
         # --- Intercambio ---
         self.add(TradeCmdSet)
+        # --- Correo ---
+        self.add(MailCmdSet)
         # --- General ---
         self.add(GeneralCmdSet)